+++ scripts/macros/2024/classFeatures/bard/lore/cuttingWords.js
@@ -1,0 +1,27 @@
+import {bardicInspiration} from '../bardicInspiration.js';
+
+
+export let cuttingWords = {
+    name: 'Cutting Words',
+    version: '1.2.2',
+    rules: 'modern',
+    config: [
+        {
+            value: 'classIdentifier',
+            label: 'CHRISPREMADES.Config.ClassIdentifier',
+            type: 'text',
+            default: 'bard',
+            category: 'homebrew',
+            homebrew: true
+        },
+        {
+            value: 'scaleIdentifier',
+            label: 'CHRISPREMADES.Config.ScaleIdentifier',
+            type: 'text',
+            default: 'bardic-inspiration',
+            category: 'homebrew',
+            homebrew: true
+        }
+    ],
+    scales: bardicInspiration.scales
+};